1차 merge

+++ src/main/java/Albaid/backend/domain/contract/application/ContractService.java
@@ -23,5 +23,5 @@ public interface ContractService {
     ResponseContractDTO updateContract(Integer contractId, RequestContractDTO request);
 
     // 경력 사항 (계약서 데이터) 가져오는 메서드
-    List<ContractDTO> getContractsForMember(Long memberId);
+    List<ContractDTO> getContractsForMember(Integer memberId);
 }
